refactor: extract games into pluggable registry architecture
