+++ src/games/color-match/index.js
@@ -1,0 +1,3 @@
+import ColorMatchGame from './ColorMatchGame';
+export const metadata = { id: 'color-match', title: 'Color Match', icon: '🎨', description: 'Stroop challenge', color: '#ff7675', pointsPerCorrect: 12 };
+export const GameComponent = ColorMatchGame;
