+++ src/games/memory-grid/index.js
@@ -1,0 +1,3 @@
+import MemoryGridGame from './MemoryGridGame';
+export const metadata = { id: 'memory-grid', title: 'Memory Grid', icon: '🧩', description: 'Remember positions', color: '#fdcb6e', pointsPerCorrect: 20 };
+export const GameComponent = MemoryGridGame;
