+++ src/games/pattern-match/index.js
@@ -1,0 +1,3 @@
+import PatternMatchGame from './PatternMatchGame';
+export const metadata = { id: 'pattern-match', title: 'Pattern Match', icon: '🔷', description: 'Find the pattern', color: '#00cec9', pointsPerCorrect: 15 };
+export const GameComponent = PatternMatchGame;
