+++ src/games/speed-math/index.js
@@ -1,0 +1,3 @@
+import SpeedMathGame from './SpeedMathGame';
+export const metadata = { id: 'speed-math', title: 'Speed Math', icon: '⚡', description: 'Solve arithmetic fast', color: '#6c5ce7', pointsPerCorrect: 10 };
+export const GameComponent = SpeedMathGame;
